+++ build.php
@@ -2,7 +2,7 @@
 /**
  * extension source version
  */
-$version = '1.6.0';
+$version = '1.6.1';
 
 /**
  * dist filename